+++ CHANGELOG.rst
@@ -42,6 +42,9 @@ Changes:
 - Added the ``structlog.stdlib.ExtraAdder`` processor that adds extra attributes of ``logging.LogRecord`` objects to the event dictionary.
   This processor can be used for adding data passed in the ``extra`` parameter of the ``logging`` module's log methods to the event dictionary.
   `#377 <https://github.com/hynek/structlog/pull/377>`_
+- Added the ``structlog.processor.CallsiteParameterAdder`` processor that adds parameters of the callsite that an event dictionary orginated from to the event dictionary.
+  This processor can be used to enrich events dictionaries with information such as the function name, line number and filename that an event dictionary orignated from.
+  `#380 <https://github.com/hynek/structlog/pull/380>`_
 
 
 ----